pyproject.toml
[tool.poetry]
name = "book-catalog"
version = "0.1.0"
description = ""
authors = ["nurbol0tt <fairtail764@gmail.com>"]
readme = "README.md"
packages = [{include = "book_catalog"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
pyjwt = "^2.8.0"
celery = "^5.4.0"
gunicorn = "^22.0.0"
redis = "^5.0.7"
python-dotenv = "^1.0.1"
python-decouple = "^3.8"
django-filter = "^24.2"
psycopg2-binary = "^2.9.9"
django-injector = "^0.3.1"
injector = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
